requirements.txt
smartapi-python
pyotp
python-dotenv
pandas
yfinance
requests
flask
scikit-learn
numpy
